deploy/docker/clickhouse-setup/requirements.txt
# OpenTelemetry core packages - using stable versions
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation>=0.42b0,<0.43b0

# OpenTelemetry exporters - matching core versions
opentelemetry-exporter-otlp>=1.21.0,<2.0.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0,<2.0.0
opentelemetry-exporter-otlp-proto-http>=1.21.0,<2.0.0

# OpenTelemetry instrumentations - all using same base version
opentelemetry-instrumentation-flask>=0.42b0,<0.43b0
opentelemetry-instrumentation-requests>=0.42b0,<0.43b0
opentelemetry-instrumentation-sqlite3>=0.42b0,<0.43b0
opentelemetry-instrumentation-psycopg2>=0.42b0,<0.43b0
opentelemetry-instrumentation-celery>=0.42b0,<0.43b0
opentelemetry-instrumentation-redis>=0.42b0,<0.43b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0,<0.43b0
opentelemetry-instrumentation-logging>=0.42b0,<0.43b0
opentelemetry-instrumentation-grpc>=0.42b0,<0.43b0

